requirements.txt
# ── Environment / Secrets ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Data Collection (Step 1) ──────────────────────────────────────────────────
requests>=2.31.0
feedparser>=6.0.11

# ── Teacher Labeling (Step 2) ─────────────────────────────────────────────────
anthropic>=0.40.0

# ── Training (Step 3) — install on Google Colab ───────────────────────────────
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
scikit-learn>=1.4.0

# ── Export + Quantize (Step 4) ────────────────────────────────────────────────
onnx>=1.16.0
onnxruntime>=1.18.0

# ── API Server (Step 5) ───────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
numpy>=1.26.0
